pyproject.toml
[tool.poetry]
name = "drf-api-permissions-postgres"
version = "0.1.0"
description = ""
authors = ["moradalkhatib <muradalkhateb@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
psycopg2-binary = "^2.9.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
